backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.3
httpx==0.27.0
python-dotenv==1.0.0
